engines/python/pyproject.toml
[project]
name = "noetic-engine"
version = "0.1.0"
description = "The reference implementation of the Noetic Engine."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "chromadb>=0.4.22",
    "pydantic>=2.5.0",
    "networkx>=3.0",
    "fastapi>=0.109.0", # For FastUI if needed, but maybe just for types
    "fastui>=0.5.0",
    "jsonpointer>=2.4",
    "json-logic-qubit>=0.0.1", # As per orchestration readme
    "langgraph>=0.0.10", # As per orchestration readme
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
